Fixed regression that prevents OAuth2 from working
On a previous commit that introduced "developer auth", a regression was
introduced that prevented OAuth2 from working (which wasn't testable at
the time because of what now seems to have been an intermittent issue on
LinkedIn's API that is now resolved.)

Also, corrected a typo in a variable name.

@ozgur - I think that this inadvertently fixes #11 and it would appear
that #2 is no longer relevant now either. I am able to work with OAuth2
with no problem. However, note that you must exchange the authorization
code for an access token very quickly (something around 5 seconds or
less from my testing) for things to work. Using the linkedin.server
example should be sufficient to demonstrate that this code works, and if
anyone is interested in seeing it in action in IPython Notebook and/or
with the help of Flask, you can check out
https://github.com/ptwobrussell/Mining-the-Social-Web-2nd-Edition/blob/master/ipynb/_AppendixB.ipynb
which is more easily viewed at
http://nbviewer.ipython.org/urls/raw.github.com/ptwobrussell/Mining-the-Social-Web-2nd-Edition/master/ipynb/_AppendixB.ipynb